fix: refactor downloadFile sample into its own file
